Minecraft/Redstone Gold Block/requirements.txt
pillow~=10.3.0
anvil-parser==0.9.0